fix: extract shared URL safety helper, protect side-panel fetch path (V-001 round 2)
依 PR #193 第二輪 reviewer 意見：

- 新增 modules/utils/urlSafety.js 共用驗證函式，offscreen.js 與
  rssManager.js 共用同一份邏輯，避免「補一個入口、漏一個入口」
- 修正 IPv6 regex 誤傷問題：fc/fd/fe80 開頭的一般網域（如
  fc.example.com、fdroid-mirror.org）不再被誤擋，改以
  hostname.includes(':') 確認為 IPv6 literal 才套用
- rssManager.js fetchRssFeed() 加入 URL 安全檢查與 redirect:'manual'，
  補上 side panel context 的 SSRF 防護缺口
- offscreen.html 改為 type="module" 載入，Makefile 對應新增 esbuild
  bundle 與 sed 步驟
- 於 urlSafety.js 頂部註明 DNS rebinding 殘餘風險為已知限制
