+++ offscreen.html
@@ -12,7 +12,7 @@
     that are not available in the Service Worker context.
     Primary use: RSS/Atom XML parsing with DOMParser.
   -->
-    <script src="offscreen.js"></script>
+    <script type="module" src="offscreen.js"></script>
 </body>
 
 </html>
